+++ scenedetect/_cli/config.py
@@ -271,6 +271,7 @@ CONFIG_MAP: ConfigDict = {
     },
     'global': {
         'backend': 'opencv',
+        'default-detector': 'detect-adaptive',
         'downscale': 0,
         'downscale-method': 'linear',
         'drop-short-scenes': False,
@@ -312,8 +313,9 @@ certain string options are stored in `CHOICE_MAP`."""
 CHOICE_MAP: Dict[str, Dict[str, List[str]]] = {
     'global': {
         'backend': ['opencv', 'pyav', 'moviepy'],
-        'verbosity': ['debug', 'info', 'warning', 'error', 'none'],
+        'default-detector': ['detect-adaptive', 'detect-content', 'detect-threshold'],
         'downscale-method': [value.name.lower() for value in Interpolation],
+        'verbosity': ['debug', 'info', 'warning', 'error', 'none'],
     },
     'split-video': {
         'preset': [